pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "quant-equity-alpha"
version = "0.1.0"
description = "Production-grade cross-sectional equity alpha platform"
authors = [{name = "Quant Team", email = "quant@example.com"}]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "pandas>=2.1.4",
    "numpy>=1.26.2",
    "polars>=0.20.2",
    "scipy>=1.11.4",
    "requests>=2.31.0",
    "pyarrow>=14.0.2",
    "scikit-learn>=1.3.2",
    "lightgbm>=4.1.0",
    "catboost>=1.2.2",
    "cvxpy>=1.4.1",
    "shap>=0.44.0",
    "plotly>=5.18.0",
    "matplotlib>=3.8.2",
    "seaborn>=0.13.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "typer>=0.9.0",
    "rich>=13.7.0",
    "tqdm>=4.66.1",
    "loguru>=0.7.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "ruff>=0.1.8",
]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --cov=src --cov-report=html --cov-report=term"
